backend/requirements.txt
fastapi
uvicorn
torch
mysql-connector
PIL
